refactor: split numbers to their own file
