+++ src/interpreter/bl_types/__init__.py
@@ -5,11 +5,12 @@
 # ruff: noqa: F401
 # flake8: noqa: F401
 from .essentials import (
-    Result, ExpressionResult, Success, Value, BLError, String, Int, Float,
-    Bool, BOOLS, Null, NULL, PythonFunction, BLFunction, Call, Class, Instance,
+    Result, ExpressionResult, Success, Value, BLError, String, Bool, BOOLS,
+    Null, NULL, PythonFunction, BLFunction, Call, Class, Instance,
     ObjectClass, ExceptionClass, NotImplementedException,
     IncorrectTypeException, VarNotFoundException, Env, Return,
     cast_to_instance,
 )
+from .numbers import Int, Float
 from .colls import BLList, BLDict, Module
 from .iterator import Item
