+++ src/interpreter/bl_types/essentials.py
@@ -354,9 +354,6 @@ class BLError(Exit, ExpressionResult):
         return self
 
 
-# endsection
-
-
 # section Values
 
 
@@ -461,6 +458,7 @@ class String(Value):
         self, other: ExpressionResult, interpreter: "ASTInterpreter",
         meta: Meta | None,
     ) -> ExpressionResult:
+        from .numbers import Int  # pylint: disable=import-outside-toplevel
         match other:
             case Int(times):
                 return String(self.value * times)
@@ -521,239 +519,6 @@ class String(Value):
         self, interpreter: "ASTInterpreter", meta: Meta | None
     ) -> "String":
         return String(f"{self.value!r}")
-
-
-# section Numbers
-
-
-@dataclass(frozen=True)
-class Int(Value):
-    """Integer type"""
-
-    value: int
-
-    @override
-    def add(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Int(other_val):
-                return Int(self.value + other_val)
-            case Float(other_val):
-                return Float(self.value + other_val)
-        return super().add(other, interpreter, meta)
-
-    @override
-    def subtract(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Int(other_val):
-                return Int(self.value - other_val)
-            case Float(other_val):
-                return Float(self.value - other_val)
-        return super().subtract(other, interpreter, meta)
-
-    @override
-    def multiply(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Int(other_val):
-                return Int(self.value * other_val)
-            case Float(other_val):
-                return Float(self.value * other_val)
-        return super().multiply(other, interpreter, meta)
-
-    @override
-    def divide(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Int(other_val) | Float(other_val):
-                try:
-                    return Float(self.value / other_val)
-                except ZeroDivisionError:
-                    return BLError(cast_to_instance(
-                        DivByZeroException.new([], interpreter, meta)
-                    ), meta)
-        return super().divide(other, interpreter, meta)
-
-    @override
-    def is_equal(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> Bool:
-        match other:
-            case Int(other_val) | Float(other_val):
-                return BOOLS[self.value == other_val]
-        return super().is_equal(other, interpreter, meta)
-
-    @override
-    def is_less(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Int(other_val) | Float(other_val):
-                return BOOLS[self.value < other_val]
-        return super().is_less(other, interpreter, meta)
-
-    @override
-    def is_less_or_equal(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Int(other_val) | Float(other_val):
-                return BOOLS[self.value <= other_val]
-        return super().is_less_or_equal(other, interpreter, meta)
-
-    @override
-    def is_greater(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Int(other_val) | Float(other_val):
-                return BOOLS[self.value > other_val]
-        return super().is_greater(other, interpreter, meta)
-
-    @override
-    def is_greater_or_equal(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Int(other_val) | Float(other_val):
-                return BOOLS[self.value >= other_val]
-        return super().is_greater_or_equal(other, interpreter, meta)
-
-    @override
-    def neg(self, interpreter: "ASTInterpreter", meta: Meta | None) -> "Int":
-        return Int(-self.value)
-
-    @override
-    def dump(self, interpreter: "ASTInterpreter", meta: Meta | None) -> String:
-        return String(repr(self.value))
-
-
-@dataclass(frozen=True)
-class Float(Value):
-    """Float type"""
-
-    value: float
-
-    @override
-    def add(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Float(other_val) | Int(other_val):
-                return Float(self.value + other_val)
-        return super().add(other, interpreter, meta)
-
-    @override
-    def subtract(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Float(other_val) | Int(other_val):
-                return Float(self.value - other_val)
-        return super().subtract(other, interpreter, meta)
-
-    @override
-    def multiply(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Float(other_val) | Int(other_val):
-                return Float(self.value * other_val)
-        return super().multiply(other, interpreter, meta)
-
-    @override
-    def divide(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Float(other_val) | Int(other_val):
-                try:
-                    return Float(self.value / other_val)
-                except ZeroDivisionError:
-                    return BLError(cast_to_instance(
-                        DivByZeroException.new([], interpreter, meta)
-                    ), meta)
-        return super().divide(other, interpreter, meta)
-
-    @override
-    def is_equal(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> Bool:
-        match other:
-            case Float(other_val) | Int(other_val):
-                return BOOLS[self.value == other_val]
-        return super().is_equal(other, interpreter, meta)
-
-    @override
-    def is_less(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Float(other_val) | Int(other_val):
-                return BOOLS[self.value < other_val]
-        return super().is_less(other, interpreter, meta)
-
-    @override
-    def is_less_or_equal(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Float(other_val) | Int(other_val):
-                return BOOLS[self.value <= other_val]
-        return super().is_less_or_equal(other, interpreter, meta)
-
-    @override
-    def is_greater(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Float(other_val) | Int(other_val):
-                return BOOLS[self.value > other_val]
-        return super().is_greater(other, interpreter, meta)
-
-    @override
-    def is_greater_or_equal(
-        self, other: ExpressionResult, interpreter: "ASTInterpreter",
-        meta: Meta | None,
-    ) -> ExpressionResult:
-        match other:
-            case Float(other_val) | Int(other_val):
-                return BOOLS[self.value >= other_val]
-        return super().is_greater_or_equal(other, interpreter, meta)
-
-    @override
-    def neg(
-        self, interpreter: "ASTInterpreter", meta: Meta | None
-    ) -> "Float":
-        return Float(-self.value)
-
-    @override
-    def dump(
-        self, interpreter: "ASTInterpreter", meta: Meta | None
-    ) -> String:
-        return String(repr(self.value))
 
 
 # section Functions
@@ -941,7 +706,6 @@ ExceptionClass = Class(String("Exception"), ObjectClass, exc_methods)
 NotImplementedException = Class(
     String("NotImplementedException"), ExceptionClass
 )
-DivByZeroException = Class(String("DivByZeroException"), ExceptionClass)
 AttrNotFoundException = Class(String("AttrNotFoundException"), ExceptionClass)
 VarNotFoundException = Class(String("VarNotFoundException"), ExceptionClass)
 IncorrectTypeException = Class(
